Refactor into image, instruments, and load_image
