+++ flows/instruments/__init__.py
@@ -1,0 +1,1 @@
+from .instruments import INSTRUMENTS, Instrument
